test(rankservice): split the commit half out, keeping the file under the cap
The reservation-expiry test pushed tests.rs to 542 lines. Split at the commit
boundary the file already had, with the shared fixtures made pub(super) rather
than duplicated.
